Se corrigieron errores varios || Se logro cambiar de color al punto en ubidots y se tienen dos dashboard, ademas se agrego un perfil de clientes y se agrego un usuario con permisos limitados

+++ video_camara.py
@@ -56,8 +56,8 @@ def func_video_camara():
 		
 		#Rescalar la imagen:
 		porcentaje = 60
-		ancho =  int(cuadro.shape[1] * deteccion.porc_video/100)
-		alto =  int(cuadro.shape[0] * deteccion.porc_video/100)
+		ancho =  int(cuadro.shape[1] * int(deteccion.porc_video)/100)
+		alto =  int(cuadro.shape[0] * int(deteccion.porc_video)/100)
 		dimensiones = (ancho, alto)
 		cuadro = cv2.resize(cuadro, dimensiones, interpolation = cv2.INTER_AREA)
